test: rename span metrics test files to mirror src and fold auto-instrumentation tests into test_instrumentor
