+++ cloudwatch-plugin-otel/tests/span_metrics/test_connector.py
